Separated db layer
