+++ db/users.db.ts
@@ -1,7 +1,7 @@
 import { getAdminFirestore } from "@/lib/firebase/firebase-admin";
 import { UserRole, UserModel } from "@/lib/auth/auth.types";
 import { Timestamp } from "firebase-admin/firestore";
-import { normalizeNotifications } from "../utils/notifications.utils";
+import { normalizeNotifications } from "../lib/utils/notifications.utils";
 
 const USERS_COLLECTION = "users";
